Simpler API for use in the relayer
